requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.22
alembic==1.12.0
pytest==7.4.3
httpx==0.25.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
openai==1.2.4
transformers==4.35.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2
python-dotenv==1.0.0
# Database drivers (uncomment as needed)
# psycopg2-binary==2.9.9  # PostgreSQL
# pymysql==1.1.0  # MySQL
# aiosqlite==0.19.0  # SQLite 